% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_cov.R
\name{data.pro}
\alias{data.pro}
\title{A sandwich covariance matrix for model parameters}
\usage{
data.pro(kk, nodes = 3)
}
\arguments{
\item{nodes}{the number of nodes required for parallel}

\item{res.list}{the list of data including the estimation of model parameters and variational parameters}
}
\value{
p_val the p-value for model parameters
}
\description{
A sandwich covariance matrix for model parameters
}
\examples{
sub.n <- 10
pos.n <- 3
vis.n <- 3
otu.n <- 100
sim_dat <- sim.data(sub.n, pos.n, vis.n, otu.n)
Y_mat <- sim_dat$Y_mat
data <- data.frame(Y=Y_mat[,1],ID=sim_dat$ID,cluster=sim_dat$cluster,x1=sim_dat$x1,x2=sim_dat$x2)
est_m <- step_alg(data,offset=NULL)
res.t <- data.pro(est_m,nodes=2)
}
