% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{sim.data}
\alias{sim.data}
\title{An approach to generate multilevel data}
\usage{
sim.data(sub.n, pos.n, vis.n, otu.n)
}
\arguments{
\item{sub.n}{the number of subjects}

\item{pos.n}{the number of clusters}

\item{vis.n}{the number of visits}

\item{otu.n}{the number of taxa}
}
\value{
data.mat  a list of data including covariates, count data
}
\description{
An approach to generate multilevel data
}
\examples{
sub.n <- 10
pos.n <- 3
vis.n <- 3
otu.n <- 100
sim_dat <- sim.data(sub.n, pos.n, vis.n, otu.n)
}
