% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_alg.R
\name{step_alg}
\alias{step_alg}
\title{An algrithm for variational parameters estimation of multilevel zero inflated negative binomial model}
\usage{
step_alg(data, trace = FALSE, offset = NULL)
}
\arguments{
\item{data}{the multilevel data}

\item{offset}{log total sequence reads}
}
\value{
beta the model paraeters for negative binomial

gamma the model paraeters for logistic regression
}
\description{
An algrithm for variational parameters estimation of multilevel zero inflated negative binomial model
}
\examples{
sub.n <- 10
pos.n <- 3
vis.n <- 3
otu.n <- 100
sim_dat <- sim.data(sub.n, pos.n, vis.n, otu.n)
Y_mat <- sim_dat$Y_mat
data <- data.frame(Y=Y_mat[,1],ID=sim_dat$ID,cluster=sim_dat$cluster,x1=sim_dat$x1,x2=sim_dat$x2)
est_m <- step_alg(data,offset=NULL)
}
